Fix data types in Floyd-Warshall implementation for better accuracy

+++ Floyd Warshall/shortest_routes.cpp
@@ -5,7 +5,7 @@ int main()
 {
     int n, e, q;
     cin >> n >> e >> q;
-    int adj_mat[n][n];
+    long long int adj_mat[n][n];
 
     for (int i = 0; i < n; i++)
     {
@@ -14,13 +14,13 @@ int main()
             if (i == j)
                 adj_mat[i][j] = 0;
             else
-                adj_mat[i][j] = INT_MAX;
+                adj_mat[i][j] = LLONG_MAX;
         }
     }
 
     while (e--)
     {
-        int a, b, w;
+        long long int a, b, w;
         cin >> a >> b >> w;
         adj_mat[a][b] = w;
         adj_mat[b][a] = w;
@@ -32,7 +32,7 @@ int main()
         {
             for (int j = 0; j < n; j++)
             {
-                if (adj_mat[i][k] != INT_MAX && adj_mat[k][j] != INT_MAX && adj_mat[i][k] + adj_mat[k][j] < adj_mat[i][j])
+                if (adj_mat[i][k] != LLONG_MAX && adj_mat[k][j] != LLONG_MAX && adj_mat[i][k] + adj_mat[k][j] < adj_mat[i][j])
                 {
                     adj_mat[i][j] = adj_mat[i][k] + adj_mat[k][j];
                 }
@@ -44,7 +44,7 @@ int main()
     {
         int s, d;
         cin >> s >> d;
-        if (adj_mat[s][d] == INT_MAX)
+        if (adj_mat[s][d] == LLONG_MAX)
             cout << -1 << endl;
         else
             cout << adj_mat[s][d] << endl;
